Move validation methods out of common and into relevant files
