+++ server/service/src/stocktake/update/validate.rs
@@ -1,11 +1,10 @@
-use repository::{RepositoryError, StocktakeLine, StocktakeRow, StorageConnection};
+use repository::{
+    EqualFilter, RepositoryError, StockLine, StockLineFilter, StockLineRepository, StocktakeLine,
+    StocktakeLineFilter, StocktakeLineRepository, StocktakeRow, StorageConnection,
+};
 
 use crate::{
-    stocktake::{
-        check_snapshot_matches_current_count, check_stock_lines_reduced_to_zero,
-        check_stocktake_exist, check_stocktake_is_not_locked, check_stocktake_not_finalised,
-        load_stocktake_lines,
-    },
+    stocktake::{check_stocktake_exist, check_stocktake_not_finalised},
     validate::check_store_id_matches,
 };
 
@@ -60,5 +59,79 @@ pub fn validate(
 impl From<RepositoryError> for UpdateStocktakeError {
     fn from(error: RepositoryError) -> Self {
         UpdateStocktakeError::DatabaseError(error)
+    }
+}
+
+fn check_snapshot_matches_current_count(
+    stocktake_lines: &[StocktakeLine],
+) -> Option<Vec<StocktakeLine>> {
+    let mut mismatches = Vec::new();
+    for line in stocktake_lines {
+        let stock_line = match &line.stock_line {
+            Some(stock_line) => stock_line,
+            None => continue,
+        };
+        if line.line.snapshot_number_of_packs != stock_line.total_number_of_packs {
+            mismatches.push(line.clone());
+        }
+    }
+    if !mismatches.is_empty() {
+        return Some(mismatches);
+    }
+    None
+}
+
+fn check_stock_lines_reduced_to_zero(
+    connection: &StorageConnection,
+    stocktake_lines: &Vec<StocktakeLine>,
+) -> Result<Option<Vec<StockLine>>, RepositoryError> {
+    let mut lines_reduced_to_zero = Vec::new();
+
+    for line in stocktake_lines {
+        let stock_line_row = match &line.stock_line {
+            Some(stock_line) => stock_line,
+            None => continue,
+        };
+        if let Some(counted_number_of_packs) = line.line.counted_number_of_packs {
+            let adjustment = stock_line_row.total_number_of_packs - counted_number_of_packs;
+
+            if adjustment > 0.0
+                && (stock_line_row.total_number_of_packs - adjustment < 0.0
+                    || stock_line_row.available_number_of_packs - adjustment < 0.0)
+            {
+                let stock_line = StockLineRepository::new(connection)
+                    .query_by_filter(
+                        StockLineFilter::new().id(EqualFilter::equal_to(&stock_line_row.id)),
+                        None,
+                    )?
+                    .pop()
+                    .ok_or(RepositoryError::NotFound)?;
+
+                lines_reduced_to_zero.push(stock_line.clone())
+            }
+        }
+    }
+
+    if !lines_reduced_to_zero.is_empty() {
+        return Ok(Some(lines_reduced_to_zero));
+    }
+    Ok(None)
+}
+
+fn load_stocktake_lines(
+    connection: &StorageConnection,
+    stocktake_id: &str,
+    store_id: &str,
+) -> Result<Vec<StocktakeLine>, RepositoryError> {
+    StocktakeLineRepository::new(connection).query_by_filter(
+        StocktakeLineFilter::new().stocktake_id(EqualFilter::equal_to(stocktake_id)),
+        Some(store_id.to_string()),
+    )
+}
+
+fn check_stocktake_is_not_locked(input: &UpdateStocktake, existing: &StocktakeRow) -> bool {
+    match &input.is_locked {
+        Some(false) => true,
+        _ => !existing.is_locked,
     }
 }
